feat(preview): add centralized preview and image loading features
This commit introduces a new PreviewPane class to encapsulate the
image preview functionality within the application. It centralizes the
handling of single-image and grid view previews, enhancing the
structure and maintainability of the code. Additionally, it implements
an ImageTaskRunner for efficient background image loading, improving
the user experience when interacting with the image viewer.
